Ensure project CFLAGS are merged with environment
CFLAGS unfortunately contain a mix of compiler and target details provided by
the user or their environment, but also flags provided by the project itself.
Particularly, -I flags to the project itself, C standard, and so on.

This uses a new variable, LIB_CFLAGS for the cflags to use when building the
library, which includes any environment provided CFLAGS.

It also includes the use of pkg-config to find the correct include paths for
libev.

Signed-off-by: Karl Palsson <karlp@tweak.net.au>