Reduce the harm caused by slow initial tunnel connections.
This change will allow multiple reconfiguring sequences to run at once,
when establishing initial tunnel connections takes an unusually long
time.

The effect of this is to repeatedly fall back and try the second-fastest
relay if connecting to the "fastest" takes a long time, as many times as
we have background "blocker" threads. This increases the odds that we'll
establish a working connection when the network is giving us a hard
time.

This change also adds more assertions and lock checks to make sure we
don't have multiple reconfiguration events in flight, outside the
specific cases defined above. This change takes advantage of the new
CONN_STATUS_CHANGING flag so one reconfig event doesn't clobber the work
being done by another.