configure.ac: fix handling of --with-*
The 'action-if-given' argument of AC_ARG_WITH is executed whenever the
--with- or --without- option is given. Setting e.g. with_tests=yes in
that branch causes the argument '--without-tests' to *enable* the tests
instead of disabling them.

In most cases, the third and fourth argument can simply be skipped
since they are optional. We only need them in the cases where we use a
different variable than with_foo, or where we want to default to yes
instead of defaulting to empty.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>