configure.ac: use pkg-config for openssl
It is better to use pkg-config to detect openssl if that is possible.
pkg-config will add e.g. -lz and -ldl when needed. If pkg-config
fails, fall back to the old approach of detecting headers and libs.
Some of the additional openssl support (e.g. adding -ldl) is moved
inside the non-pkg-config path.

Since AC_CHECK_LIBS adds the library found to LIBS, do the same in
the pkg-config case. Normally the Makefile.am should instead use
OPENSSL_LIBS where needed, but this is not done consistently.

When --with-openssl=DIR is given, still perform the test (both with
pkg-config and by checking headers and libs). I.e., remove
$specialssldir.

While we're at it, simplify the headers checks by merging them into a
single AC_CHECK_HEADERS.

Note that it is (still) not an error when openssl is not found,
although the build will then fail.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>