Makefile.am: use $(top_builddir) for LDADD
When building out-of-tree, top_builddir != top_srcdir. For includes,
top_srcdir should be used because they are not generated. For binaries,
however (i.e. the .la files included in LDADD), top_builddir must be
used. Otherwise linking will fail because the .la file is not found in
the source directory.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>