Add flags to tell DDNS whether we want IPv4 or IPv6 DNS records (or both).
The underlying assumption here, is that many/most public relays will be on
both networks and the dynamic DNS server knows how to automatically map one
type of address to another.

This makes it possible to express a desire for visibility on a network
that differs from the one we connect over; so folks with only IPv4 can
still be visible on IPv6 (or vice versa).