Fix to work ok with 64 bit time_t (time64)
In several places, time_t arguments were formatted using the "%d"
format specifier, which is not compatible with 64bit time_t
(as for example in musl >=1.2.0)

This commit now uses "%lld" with an explicit cast of the
arguments to long long. The cast ensures that the code is still
safe to compile with 32 bit time_t (but introduces a slight
performance penalty in that case).